One of the limitation in creating a Git commit description is that it is limited to 50 characters, which sometimes is not much for a breakthrough changes in our code. To accommodate that, we can use a better way to commit and create the commit message. Creating Commit with Description Usually, we will create a commit by using git commit -m "<some message here>". There is a better way to create commit message, that is by using git commit (no other parameter). To give a better visualization, let me give you an example. I’m already making changes in my README.md file, so now we can add the changes to the staging area using git add . (or using git add <file-name> if you prefer). Then after that, we can use git commit. You will be shown a console text editor called Vim. Don’t panic, we only need to use several basic usages of Vim. To change the mode into insert mode, we will need to press the ‘I’ character in our keyboard. Then notice that the lower-left corner will show text ‘-- INSERT --’, that means that now we can write our commit message.
One of the limitation in creating a Git commit description is that it is limited to 50 characters, which sometimes is not much for a breakthrough changes in our code. To accommodate that, we can use a better way to commit and create the commit message.
Creating Commit with Description
Usually, we will create a commit by using git commit -m "<some message here>". There is a better way to create commit message, that is by using git commit (no other parameter). To give a better visualization, let me give you an example.
I’m already making changes in my README.md file, so now we can add the changes to the staging area using git add . (or using git add <file-name> if you prefer). Then after that, we can use git commit.
You will be shown a console text editor called Vim. Don’t panic, we only need to use several basic usages of Vim. To change the mode into insert mode, we will need to press the ‘I’ character in our keyboard. Then notice that the lower-left corner will show text ‘-- INSERT --’, that means that now we can write our commit message.